# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PostEntity and Comments API
  description: |
    RESTful API for managing posts, comments, post details, and tags.
    
    This API allows you to:
    - Create, read, update, and delete posts
    - Add comments to posts
    - Manage post details and tags
    - Associate tags with posts
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/tags:
    get:
      tags:
        - Tags
      summary: Get all tagEntities
      description: Retrieves a list of all tagEntities
      operationId: getAllTags
      responses:
        '200':
          description: Successfully retrieved tagEntities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
              example:
                - id: 1
                  name: "Spring Boot"
                - id: 2
                  name: "Java"

    post:
      tags:
        - Tags
      summary: Create a new tag
      description: Creates a new tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
            example:
              name: "Spring Boot"
      responses:
        '201':
          description: TagEntity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: TagEntity with this name already exists
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tags/{id}:
    get:
      tags:
        - Tags
      summary: Get tag by ID
      description: Retrieves a specific tag by its ID
      operationId: getTagById
      parameters:
        - name: id
          in: path
          required: true
          description: TagEntity ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: TagEntity not found

    put:
      tags:
        - Tags
      summary: Update a tag
      description: Updates an existing tag
      operationId: updateTag
      parameters:
        - name: id
          in: path
          required: true
          description: TagEntity ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '200':
          description: TagEntity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: TagEntity not found
        '409':
          description: TagEntity with this name already exists

    delete:
      tags:
        - Tags
      summary: Delete a tag
      description: Deletes a tag by its ID
      operationId: deleteTag
      parameters:
        - name: id
          in: path
          required: true
          description: TagEntity ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: TagEntity deleted successfully
        '404':
          description: TagEntity not found
  /api/tags/name/{name}:
    get:
      tags:
        - Tags
      summary: Get tag by name
      description: Retrieves a tag by its name
      operationId: getTagByName
      parameters:
        - name: name
          in: path
          required: true
          description: TagEntity name
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: TagEntity not found
components:
...
  schemas:
...
    Tag:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the tag
          example: 1
        name:
          type: string
          description: The name of the tag (must be unique)
          example: "Spring Boot"
...
    Error:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: "Bad Request"
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: string
          example: ["Title is required"]
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/api/postEntities"
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 400
            message: "Bad Request"
            success: false
            errors: ["Title is required"]
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 500
            message: "Internal Server Error"
            success: false
            errors: ["An unexpected error occurred"]